Update Contact section
* Break out SocialIcon to it's own component
* Adjust test to not use snapshot
