+++ src/components/contact/social-icon.js
@@ -1,0 +1,28 @@
+// @flow
+import * as React from 'react';
+import { css } from 'emotion';
+
+import themeUtils from '../themeUtils';
+
+type IconProps = {
+  link: string,
+  component: React.Node,
+};
+
+const socialLink = css`
+  color: ${themeUtils.mediumAccent};
+  margin-right: 0.5rem;
+  transition: color 800ms;
+
+  &:hover {
+    color: ${themeUtils.complementaryDark};
+  }
+`;
+
+const SocialIcon = ({ component, link }: IconProps) => (
+  <a className={socialLink} href={link}>
+    {component}
+  </a>
+);
+
+export default SocialIcon;
